Rename module to method
